Improved LTI connections (#707)
* Preset db course renaming to follow cohort structure

* LTI launch url make use of urlparse

* Removed unused Assignment model comment property

* Improved group creation response when the name is already in use.

* Gitignore expanded to include psql dump

* Added db dump and restore make commands

* Factory updates

* Removed some unused statements from Serializer

* Migration and lti_ids model changes

* Import statement order

* Reintroduced get_lti_linked

* Rework to active_lti_id

* Corrected migration chain

* Added tests for the lti_id_set functionality

* Finalised data structure

* Processing "lti_id"

* Moved publish_all_assignment_grades call

* Further processing of lti_id replacement

* Finished processing sourcedid search

* Clear grade_url and sourcedid for all assignment's journals on save

* Clear grade_url and sourcedid for all assignment's journals on save

* Fixed test and assignment save logic

* Added LTI multi-course tests

* Added get lti linked field to journal serializer

* Removed OLD_ASSIGNMENT state

* Added needs_lti_link to journal serializer

* Course assignment_lti_id_set start

* Move feedback button to bottom of page, regardless of page height

* Changed loading spinner upon launching via LTI

* Needs lti link checks in progress

* Entry  tests added

* Update journal grade upon correct lti id

* Improved docstring

* Tests added for setting the sourcedid and grade_url

* Assignment lti course links in serializer

* Course assignment_lti_id_set logic, tests need fixing

* Added LTI tests

* Updated LTI launch and login flow

* Migration update

* Removed v-else from template

* Improved get active course function to actually return a course

* Removed outdated watcher

* Removed star icon for bonus points

* Added LTI warning banner to journal pages

* Added dropdown for changing active LTI link

* Only allow removal from inactive LTI assignments

* Fixed linter issues

* Migration fix

* Reintroduced lti_ids model untill migration finishes

* AttributeError fix guess

* Removed line break

* Fixed console error for non-lti assignments

* Don't allow journal edits until visiting the LTI assignment front end

* Reversed filter

* Fixed frontend entryGradePublished bug

* Fixed linter errors

* Removed auto migration

* removed lti_ids

* Check if assignment is already uses in course

* Fixed migration

* removed unnecessary code

+++ src/django/VLE/views/entry.py
@@ -50,6 +50,9 @@ class EntryView(viewsets.ViewSet):
         if assignment.is_locked():
             return response.forbidden('The assignment is locked, entries can no longer be edited/changed.')
 
+        if journal.needs_lti_link():
+            return response.forbidden(journal.outdated_link_warning_msg)
+
         # Check if the template is available
         if not (node_id or assignment.format.template_set.filter(archived=False, preset_only=False,
                                                                  pk=template.pk).exists()):
@@ -66,8 +69,7 @@ class EntryView(viewsets.ViewSet):
             node = factory.make_node(journal, entry)
 
         # Notify teacher on new entry
-        # TODO Verbose name for check
-        if not (node.journal.sourcedid is None or node.entry.vle_coupling != Entry.NEED_SUBMISSION):
+        if (node.journal.sourcedid and node.entry.vle_coupling == Entry.NEED_SUBMISSION):
             lti_tasks.needs_grading.delay(node.pk)
 
         for content in content_list:
@@ -129,6 +131,8 @@ class EntryView(viewsets.ViewSet):
             return response.bad_request('You are not allowed to edit graded entries.')
         if entry.is_locked():
             return response.bad_request('You are not allowed to edit locked entries.')
+        if journal.needs_lti_link():
+            return response.forbidden(journal.outdated_link_warning_msg)
 
         # Check for required fields
         entry_utils.check_required_fields(entry.template, content_list)
@@ -188,9 +192,10 @@ class EntryView(viewsets.ViewSet):
                 return response.forbidden('You are not allowed to delete locked entries.')
             if assignment.is_locked():
                 return response.forbidden('You are not allowed to delete entries in a locked assignment.')
-
         elif not request.user.is_superuser:
             return response.forbidden('You are not allowed to delete someone else\'s entry.')
+        if journal.needs_lti_link():
+            return response.forbidden(journal.outdated_link_warning_msg)
 
         if entry.node.type != Node.ENTRYDEADLINE:
             entry.node.delete()
